divided code into smaller components (emotionbutton, subemotionbutton)
